Spliting components in different files
