+++ src/02-component-patterns/components/ProductCard.tsx
@@ -1,68 +1,11 @@
-import { createContext, ReactElement, useContext } from 'react';
+import { createContext } from 'react';
 import { useProduct } from '../hooks/useProduct';
-import noImage from '../assets/no-image.jpg';
+import { ProductCardProps, ProductContextProps } from '../interfaces/interfaces';
 import styles from '../styles/styles.module.css';
 
-interface ProductContextProps {
-	counter: number;
-	increaseBy: (value: number) => void;
-	product: Product;
-}
-
-const ProductContext = createContext({} as ProductContextProps);
+export const ProductContext = createContext({} as ProductContextProps);
 
 const { Provider } = ProductContext;
-
-interface ProductCardProps {
-	children?: ReactElement | ReactElement[];
-	product: Product;
-}
-
-interface Product {
-	id: string;
-	title: string;
-	img?: string;
-}
-
-export const ProductImage = ({ img = '' }) => {
-	const { product } = useContext(ProductContext);
-
-	let imgToShow: string;
-
-	if (img) {
-		imgToShow = img;
-	} else if (product.img) {
-		imgToShow = product.img;
-	} else {
-		imgToShow = noImage;
-	}
-
-	return <img className={styles.productImg} src={imgToShow} alt='Product' />;
-};
-
-export const ProductTitle = ({ title }: { title?: string }) => {
-	const { product } = useContext(ProductContext);
-
-	return <span className={styles.productDescription}>{title ? title : product.title}</span>;
-};
-
-export const ProductButtons = () => {
-	const { counter, increaseBy } = useContext(ProductContext);
-
-	return (
-		<div className={styles.buttonsContainer}>
-			<button className={styles.buttonMinus} onClick={() => increaseBy(-1)}>
-				-
-			</button>
-
-			<div className={styles.countLabel}>{counter}</div>
-
-			<button className={styles.buttonAdd} onClick={() => increaseBy(1)}>
-				+
-			</button>
-		</div>
-	);
-};
 
 export const ProductCard = ({ children, product }: ProductCardProps) => {
 	const { counter, increaseBy } = useProduct();
@@ -73,7 +16,3 @@ export const ProductCard = ({ children, product }: ProductCardProps) => {
 		</Provider>
 	);
 };
-
-ProductCard.Title = ProductTitle;
-ProductCard.Image = ProductImage;
-ProductCard.Buttons = ProductButtons;
