+++ src/02-component-patterns/components/index.ts
@@ -1,0 +1,4 @@
+export { ProductButtons } from './ProductButtons';
+export { ProductCard } from './ProductCard';
+export { ProductImage } from './ProductImage';
+export { ProductTitle } from './ProductTitle';
