+++ src/02-component-patterns/interfaces/interfaces.ts
@@ -1,0 +1,18 @@
+import { ReactElement } from 'react';
+
+export interface ProductContextProps {
+	counter: number;
+	increaseBy: (value: number) => void;
+	product: Product;
+}
+
+export interface ProductCardProps {
+	children?: ReactElement | ReactElement[];
+	product: Product;
+}
+
+export interface Product {
+	id: string;
+	title: string;
+	img?: string;
+}
