+++ src/02-component-patterns/pages/ShoppingPage.tsx
@@ -1,4 +1,4 @@
-import { ProductButtons, ProductCard, ProductImage, ProductTitle } from '../components/ProductCard';
+import { ProductButtons, ProductCard, ProductImage, ProductTitle } from '../components';
 
 const product = {
 	id: '1',
